Edit values
Fixed the value of the government legislation against discrimination, because the libertarian response is that there should not be any government legislation of that. Currently agreement with that statement amounts to making one more libertarian.

As for the military question it isn't strictly answerable. The answer to whether a strong military is needed is to ponder "needed to what end and under what circumstances?" If it is the case that you do not want to be molested by foreigners and those foreigners have a well equipped military then a strong military to fight them seems like it may well be needed. A priori we cannot know this though. Decided that if I was to remove a question I might as well add one in its place, hence the secession question.